requirements.txt
appdirs==1.4.4
asgiref==3.3.4
astroid==2.5.2
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2
django-appconf==1.0.4
django-compressor==2.4
django-libsass==0.8
filelock==3.0.12
isort==5.8.0
lazy-object-proxy==1.6.0
libsass==0.20.1
mccabe==0.6.1
psycopg2==2.8.6
pylint==2.7.2
pylint-django==2.4.2
pylint-plugin-utils==0.6
pytz==2021.1
PyYAML==5.4.1
rcssmin==1.0.6
rjsmin==1.1.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.7.4.3
virtualenv==20.4.3
whitenoise==5.2.0
wrapt==1.12.1
